feat(auth): Add endpoint to get enhanced user profile
